Move openslide-hash.h into openslide-private.h
struct _openslide_hash is needed by openslide-private.h, so everything in
the library ends up including openslide-hash.h anyway and there's no point
in maintaining a separate header file.
